pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
]
